+++ src/main/java/net/jp/hellparadise/retake/components/RetakeButton.java
@@ -1,0 +1,9 @@
+package net.jp.hellparadise.retake.components;
+
+import net.minecraft.client.gui.GuiButton;
+
+public class RetakeButton extends GuiButton {
+    public RetakeButton(int buttonId, int x, int y, String buttonText) {
+        super(buttonId, x, y, buttonText);
+    }
+}