+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.1] - 2024-01-01
+
+### Added
+- New button to death gui.
+- Implementation of config.
+- Cooldown logic (cooldown info stored on server side, client will queue the info via packet).
+
+### Changed
+- Adjust exist buttons height to inject new button.
+
 ## [0.0] - 2024-01-01
 
 ### Added